+++ R/counties.R
@@ -26,6 +26,7 @@ redist.county.relabel <- function(adj, counties,  simplify = TRUE) {
     if ("numeric" %in% class(counties) || "integer" %in% class(counties)) {
         counties <- as.character(counties)
     }
+
     sorted <- sort(unique(counties))
     groups <- match(counties, sorted)